Add files via upload
The composition of the Nifty-50 index is periodically reviewed and may change based on the market capitalization and trading activity of the listed companies. It includes companies from various sectors, giving investors a diverse view of the Indian economy.

Keep in mind that financial markets are constantly changing, and new information may be available beyond my last update. To get the most up-to-date information on the Nifty-50 index, I recommend checking a reliable financial news source or the official website of the National Stock Exchange of India.